requirements/base.txt
Django>=3.0
psycopg2-binary>=2.8
requests_html==0.10.0
fake-useragent==0.1.11
djangorestframework==3.12.2
Markdown==3.3.3
django-filter==2.4.0
